+++ week3/include/randarray.h
@@ -1,0 +1,15 @@
+/* Copyright 2021 Carlos Pinto Machado */
+/* See LICENSE for details */
+
+#ifndef WEEK2_INCLUDE_RANDARRAY_H_
+#define WEEK2_INCLUDE_RANDARRAY_H_
+
+/* FUNCTION DECLARATIONS */
+
+/* randArrayLong: Generate an array of a given length with random integers */
+long* randArrayLong(long len, long topLimit);
+
+/* printDivisible: prints results divisible by 2, 3, 5, 7 */
+void printDivisible(long len, long* array);
+
+#endif /* WEEK2_INCLUDE_RANDARRAY_H_ */